+++ tests/pipeline/test_base.py
@@ -1,11 +1,8 @@
-from collections import defaultdict
-from collections.abc import Generator
-from typing import Any, Optional
 from unittest.mock import MagicMock, patch
 
+import pytest
 from dagster import (
     AssetKey,
-    AssetsDefinition,
     DagsterInstance,
     RunRequest,
     SkipReason,
@@ -13,7 +10,7 @@ from dagster import (
     build_sensor_context,
 )
 from dagster._core.test_utils import instance_for_test
-import pytest
+from pytest import MonkeyPatch
 
 from mex.extractors.pipeline.base import load_job_definitions, monitor_jobs_sensor
 from mex.extractors.pipeline.checks.models.check import AssetCheck, AssetCheckRule
@@ -96,77 +93,70 @@ def test_monitor_triggers_if_new_jobs_finished() -> None:
             assert result.run_key == "1100"
 
 
-# @pytest.mark.parametrize(
-#     (
-#         "asset_metadata",
-#         "rules",
-#         "expected_check_created",
-#     ),
-#     [
-#         # Single asset with one x_items_more_than rule
-#         (
-#             {"entity_type": "user"},
-#             [{"fail_if": "x_items_more_than", "value": 5, "time_frame": "7d"}],
-#             True,
-#         ),
-#         # Single asset with one x_items_less_than rule
-#         (
-#             {"entity_type": "order"},
-#             [{"fail_if": "x_items_less_than", "value": 10, "time_frame": "7d"}],
-#             True,
-#         ),
-#         # Single asset with no entity_type → no checks
-#         (
-#             {},
-#             [{"fail_if": "x_items_more_than", "value": 5, "time_frame": "7d"}],
-#             False,
-#         ),
-#         # Single asset with multiple rules
-#         (
-#             {"entity_type": "payment"},
-#             [
-#                 {"fail_if": "x_items_more_than", "value": 5, "time_frame": "7d"},
-#                 {"fail_if": "x_items_less_than", "value": 2, "time_frame": "7d"},
-#             ],
-#             True,
-#         ),
-#     ],
-# )
-def test_asset_checks_created(
-    monkeypatch: pytest.MonkeyPatch,
-    # asset_metadata: dict[str, str],
-    # rules: dict,
-    # *,
-    # expected_check_created: bool,
+@pytest.mark.parametrize(
+    (
+        "group_name",
+        "metadata",
+        "return_value",
+        "rules",
+        "expected_check_created",
+    ),
+    [
+        (
+            "blueant",
+            {"entity_type": "activity"},
+            1,
+            [{"fail_if": "x_items_more_than", "value": 5, "time_frame": "7d"}],
+            1,
+        ),
+        (
+            "blueant",
+            {},
+            2,
+            [{"fail_if": "x_items_more_than", "value": 5, "time_frame": "7d"}],
+            0,
+        ),
+        (
+            "international_projects",
+            {"entity_type": "activity"},
+            3,
+            [
+                {"fail_if": "x_items_more_than", "value": 20, "time_frame": "7d"},
+                {"fail_if": "x_items_less_than", "value": 5, "time_frame": "7d"},
+            ],
+            2,
+        ),
+    ],
+    ids=["one_check_created", "no_check_created", "two_checks_created"],
+)
+def test_asset_checks_created(  # noqa: PLR0913
+    monkeypatch: MonkeyPatch,
+    group_name: str,
+    metadata: dict[str, str],
+    return_value: int,
+    rules: list[dict[str, str]],
+    expected_check_created: int,
 ) -> None:
-    @asset(group_name="group_1", metadata={"entity_type": "activity"})
-    def asset_with_entity()-> int:
-            return 1
+    @asset(key=AssetKey(["test_asset"]), group_name=group_name, metadata=metadata)
+    def test_asset() -> int:
+        return return_value
 
-        # Asset without entity_type → should not create checks
-    @asset(group_name="group_2", metadata={})
-    def asset_without_entity()-> int:
-        return 2
-
-    mocked_assets = [asset_with_entity, asset_without_entity]
     monkeypatch.setattr(
         "mex.extractors.pipeline.base.load_assets_from_package_module",
-        lambda module: mocked_assets,
+        lambda module: [test_asset],
     )
-    load_calls = []
-    def load_asset_check_mock(group_name, entity_name) -> AssetCheck:
-        load_calls.append(entity_name)
-        return AssetCheck(
-            rules=[AssetCheckRule(fail_if="x_items_more_than", value=5, time_frame="7d")]
-        )
-    # check_model = AssetCheck(rules=[AssetCheckRule(**rules)])
 
-    # monkeypatch.setattr(
-    #     "mex.extractors.pipeline.base.load_asset_check_from_settings",
-    #     lambda *args, **kwargs: check_model,
-    # )
+    load_calls = []
+
+    def load_asset_check_mock(group_name: str, entity_name: str) -> AssetCheck:  # noqa: ARG001
+        load_calls.append(entity_name)
+        return AssetCheck(rules=[AssetCheckRule(**rule) for rule in rules])
+
+    monkeypatch.setattr(
+        "mex.extractors.pipeline.base.load_asset_check_from_settings",
+        load_asset_check_mock,
+    )
 
     defs = load_job_definitions()
 
-    checks_created = len(defs.asset_checks) > 0
-    assert checks_created
+    assert len(list(defs.asset_checks or [])) == expected_check_created